requirements.txt
# UI
streamlit

# scraping
google-search-results

# llama_index
llama_index
llama-index-llms-anthropic
chromadb
llama-index-vector-stores-chroma
llama-index-embeddings-huggingface

# langchain
langchain
langchain-community
langchain-text-splitters
unstructured
langchain-anthropic

# eval
ragas
transformers
sentence-transformers
# spacy

# miscellaneous
python-dotenv
uuid